Fix Bug with Agent Data
In some alert cases, the agent wont have an IP. To prevent issues like this in the future, the agent string creation has been converted to dictionary get methods.

Not sure if any of the other fields would actually have this issue but why not do it for all if we're doing it for one.