Refactor Discord Field Creation
This commit adds a function to the base handler to do most of the work of discord field creation in a single place. Instead of doing deep copies and changing fields, the function provides an interface for each handler to provide a name and value and it will do the deep copies and fill the entries.

This reduces a huge amount of duplicate code and also provides some protection against value errors on the Discord API side because of the type checking on the function.