Patch Bug in SuricataEve Handler
The Suricata eve handler had a bug with the DNS parsing that would cause exceptions. This was due to the DNS queries being wrapped in a list.

I'm not aware of a case where there would be more than a 1 element list so currently using the 0th index.

Also made improvements to exception handling chasing down this issue.