Made Test Suite Generic
This commit restructures the test suite to require much less boilerplate for adding new test cases for the handlers.

Test cases for the handlers can now be added with zero additional code.

Instead, the test code expects a directory in test/rule_handlers with the same name as the .py file the handler is defined in. In that directory should be 3 files:
- input.json
- expected_description.json
- expected_fields.json

These files describe the inputs and expected outputs of the handler.