Testing Suite for Handlers (#1)
* Begin Building Testing

This commit adds very preliminary testing for handlers.

It does so using pytest. Each handler will get its own directory with the test code and input data. This should allow for code commonality between handler tests.

* Add Requests Dependency

* Refactor Alert ID

Changes what was previously known as alert_id to rule_id and adds the actual alert_id to all alerts.

* Made Test Suite Generic

This commit restructures the test suite to require much less boilerplate for adding new test cases for the handlers.

Test cases for the handlers can now be added with zero additional code.

Instead, the test code expects a directory in test/rule_handlers with the same name as the .py file the handler is defined in. In that directory should be 3 files:
- input.json
- expected_description.json
- expected_fields.json

These files describe the inputs and expected outputs of the handler.

* Add Tests For More Handlers

* Remove debug code from test_all_handlers.py

* Add Test Cases for Remaining Handlers

* Add Script for Generating Test Case Templates