Add SSH Auth Handler and Base Class GeoLocation String Function
This commit adds an SSH auth handler that should be able to handle
a few cases where SSH alerts are triggered.

As part of that work, a more unified way of handling GeoLocation
blocks from Wazuh was needed to a function to do so was
added to the base_handler.